+++ package/index.ts
@@ -14,13 +14,14 @@ import {
   UserThemeContext
 } from "./config/contextHandler";
 import ConfigProvider from "./ConfigProvider";
+import { updateModuleContextData } from "./context/wrappidAction";
 import WrappidContext, {
   resetWrappidContext,
   updateWrappidContext,
   WrappidDataContext,
   WrappidDispatchContext
 } from "./context/WrappidContext";
-import { UPDATE_DEFAULT_THEME, UPDATE_DEVELOPMENT_DATA, UPDATE_PAGE_THEME } from "./context/wrappidReducer";
+import { UPDATE_DEFAULT_THEME, UPDATE_DEVELOPMENT_DATA, UPDATE_MODULE_DATA, UPDATE_PAGE_THEME } from "./context/wrappidReducer";
 import WrappidSyncer, { WrappidData } from "./context/WrappidSyncer";
 import StyledComponentsClasses from "./styledComponents/StyledComponentsClasses";
 import StylesProvider, { theme } from "./StylesProvider";
@@ -48,9 +49,7 @@ export {
   /**
    * Style Classes Constants
    */
-  StyledComponentsClasses, StylesProvider, StylesThemesContext, theme, ThemeContext,
-  UPDATE_DEFAULT_THEME, UPDATE_DEVELOPMENT_DATA, UPDATE_PAGE_THEME,
-  updateWrappidContext, UserThemeContext,
+  StyledComponentsClasses, StylesProvider, StylesThemesContext, theme, ThemeContext, UPDATE_DEFAULT_THEME, UPDATE_DEVELOPMENT_DATA, UPDATE_MODULE_DATA, UPDATE_PAGE_THEME, updateModuleContextData, updateWrappidContext, UserThemeContext,
   /**
    * Utility styles
    */