+++ CHANGELOG.md
@@ -2,6 +2,28 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.0.139](https://github.com/wrappid/styles/compare/v0.0.138...v0.0.139) (2024-03-21)
+
+
+### Bug Fixes
+
+* **core:** :bug: bug fixed for bg dotted background styles and a text styles css ([caec533](https://github.com/wrappid/styles/commit/caec53349348798a93ec050b912299398d5a6d9f)), closes [#72](https://github.com/wrappid/styles/issues/72)
+
+### [0.0.138](https://github.com/wrappid/styles/compare/v0.0.137...v0.0.138) (2024-03-20)
+
+
+### Features
+
+* **core:** :sparkles: created styles for transform rotate and a background ([35125c7](https://github.com/wrappid/styles/commit/35125c71bb873c88314d65222755735339b7a5b0)), closes [#72](https://github.com/wrappid/styles/issues/72)
+
+### 0.0.137 (2024-03-20)
+
+### 0.0.136 (2024-03-20)
+
+### 0.0.135 (2024-03-19)
+
+### 0.0.134 (2024-03-18)
+
 ### 0.0.133 (2024-03-16)
 
 ### 0.0.132 (2024-03-16)